Update `size_in_bytes` to `FileSize` (#27)
This aligns nicely with `FileMeta`:

https://github.com/dl-rs-private/delta-kernel-rs/blob/fd77665e9215e26dc74b1d74c42b67e11ebe38db/kernel/src/lib.rs#L190-L198

+++ kernel/src/transaction/mod.rs
@@ -380,7 +380,7 @@ impl Transaction {
 
             let content_root_action = ContentRoot {
                 path: file_meta.location.to_string(),
-                size_in_bytes: file_meta.size as i64,
+                size_in_bytes: file_meta.size,
             };
 
             // Use the log schema to wrap ContentRoot in a "contentRoot" field
